Update SyPGP references and moved to internal package location
