+++ cmd/internal/cli/key_search.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/hpcng/singularity/docs"
 	"github.com/hpcng/singularity/internal/pkg/remote/endpoint"
+	"github.com/hpcng/singularity/internal/pkg/sypgp"
 	"github.com/hpcng/singularity/pkg/sylog"
-	"github.com/hpcng/singularity/pkg/sypgp"
 	"github.com/spf13/cobra"
 	"github.com/sylabs/scs-key-client/client"
 )
