+++ internal/app/singularity/sign.go
@@ -8,7 +8,7 @@ package singularity
 import (
 	"github.com/hpcng/sif/pkg/integrity"
 	"github.com/hpcng/sif/pkg/sif"
-	"github.com/hpcng/singularity/pkg/sypgp"
+	"github.com/hpcng/singularity/internal/pkg/sypgp"
 )
 
 type signer struct {
