+++ internal/app/singularity/verify.go
@@ -15,7 +15,7 @@ import (
 	"github.com/hpcng/sif/pkg/integrity"
 	"github.com/hpcng/sif/pkg/sif"
 	"github.com/hpcng/singularity/internal/pkg/buildcfg"
-	"github.com/hpcng/singularity/pkg/sypgp"
+	"github.com/hpcng/singularity/internal/pkg/sypgp"
 	"github.com/sylabs/scs-key-client/client"
 	"golang.org/x/crypto/openpgp"
 )
