+++ internal/pkg/runtime/engine/singularity/prepare_linux.go
@@ -27,6 +27,7 @@ import (
 	"github.com/hpcng/singularity/internal/pkg/security"
 	"github.com/hpcng/singularity/internal/pkg/security/seccomp"
 	"github.com/hpcng/singularity/internal/pkg/syecl"
+	"github.com/hpcng/singularity/internal/pkg/sypgp"
 	"github.com/hpcng/singularity/internal/pkg/util/fs"
 	"github.com/hpcng/singularity/internal/pkg/util/fs/overlay"
 	"github.com/hpcng/singularity/internal/pkg/util/mainthread"
@@ -36,7 +37,6 @@ import (
 	"github.com/hpcng/singularity/pkg/runtime/engine/config"
 	singularityConfig "github.com/hpcng/singularity/pkg/runtime/engine/singularity/config"
 	"github.com/hpcng/singularity/pkg/sylog"
-	"github.com/hpcng/singularity/pkg/sypgp"
 	"github.com/hpcng/singularity/pkg/util/capabilities"
 	"github.com/hpcng/singularity/pkg/util/fs/proc"
 	"github.com/hpcng/singularity/pkg/util/namespaces"
